refactor: move plan ownership from team to session
Plan was stored on Team and only built during import, causing stale
plans when team composition or member specs changed through other
operations. By making Session own the plan and building it fresh at
session start, the staleness problem is eliminated by design.

- Remove Plan field from domain.Team and teams DB table
- Add Plan field to domain.Session and sessions DB table
- Move BuildPlan logic and all helpers (auth, command, config, prompt)
  from team service to session service
- Session.Start() now builds a fresh plan from current team state
- Session.Send() uses session.Plan for name lookup instead of team.Plan
- Remove BuildPlan call from import command

Closes #18
