+++ internal/app/session/auth_test.go
@@ -1,4 +1,4 @@
-package team
+package session
 
 import (
 	"testing"
