+++ internal/app/session/command.go
@@ -1,4 +1,4 @@
-package team
+package session
 
 import (
 	"fmt"
