+++ internal/app/session/config_test.go
@@ -1,4 +1,4 @@
-package team
+package session
 
 import (
 	"encoding/json"
